feat/ adicionando todas logicas do processo de venda, controle de pagamento e controle do estoque do supermercado

+++ super-mercado/src/main/java/org/example/entity/Jornada.java
@@ -2,6 +2,7 @@ package org.example.entity;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -12,8 +13,8 @@ public class Jornada {
     private Long id;
     private LocalDateTime horarioAbertura;
     private LocalDateTime horarioFechamento;
-    private Double valorInicial;
-    private Double valorFinal;
+    private BigDecimal valorInicial;
+    private BigDecimal valorFinal;
     @ManyToOne
     @JoinColumn(name = "id_caixa")
     private Caixa caixa;
@@ -21,7 +22,7 @@ public class Jornada {
     @JoinColumn(name = "id_funcionario")
     private Funcionario funcionario;
 
-    public Jornada(LocalDateTime horarioAbertura, Double valorInicial, Caixa caixa, Funcionario funcionario) {
+    public Jornada(LocalDateTime horarioAbertura, BigDecimal valorInicial, Caixa caixa, Funcionario funcionario) {
         this.horarioAbertura = horarioAbertura;
         this.valorInicial = valorInicial;
         this.caixa = caixa;
@@ -55,19 +56,19 @@ public class Jornada {
         this.horarioFechamento = horarioFechamento;
     }
 
-    public Double getValorInicial() {
+    public BigDecimal getValorInicial() {
         return valorInicial;
     }
 
-    public void setValorInicial(Double valorInicial) {
+    public void setValorInicial(BigDecimal valorInicial) {
         this.valorInicial = valorInicial;
     }
 
-    public Double getValorFinal() {
+    public BigDecimal getValorFinal() {
         return valorFinal;
     }
 
-    public void setValorFinal(Double valorFinal) {
+    public void setValorFinal(BigDecimal valorFinal) {
         this.valorFinal = valorFinal;
     }
